complementary filter: K is float
And I wondered why complementary filters have not been working for a year

+++ include/behaviour_control/complementary_filter.h
@@ -7,6 +7,6 @@ class ComplementaryFilter{
         ComplementaryFilter(float k);
         float calculate(float f);
     private:
-        int K;
+        float K;
         float oldVal;
 };
